# api/swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.0.0",
        "title": "League of Legends Scouting API",
        "description": "Gathers information on League of Legend players to scout out picks and bans.",
        "license": {}
    },
    "basePath": "/",
    "tags": [],
    "schemes": [
        "https"
    ],
    "produces": [
        "application/json"
    ],
    "paths": {
        "/team": {
            "post": {
                "tags": [
                    "Team"
                ],
                "summary": "Adds a team to the database and returns the values",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "type": "string",
                        "required": true,
                        "x-example": "Redpanda0129",
                        "description": "Player's Username."
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Player"
                        }
                    }
                }
            }
        },
        "/player": {
            "get": {
                "tags": [
                    "Player"
                ],
                "summary": "Finds the information of a player by their summoner(player) name and returns stats for that player's in game information.",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "type": "string",
                        "required": true,
                        "x-example": "Redpanda0129",
                        "description": "Player's Username."
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Player"
                        }
                    }
                }
            }
        },
        "/matches": {
            "get": {
                "tags": [
                    "Matches"
                ],
                "summary": "Finds the information of a player by their summoner(player) name and returns stats for that player from his last game.",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "type": "string",
                        "required": true,
                        "x-example": "Redpanda0129",
                        "description": "Player's Username."
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Match"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "Player": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string",
                    "uniqueItems": true
                },
                "puuid": {
                    "type": "string",
                    "uniqueItems": true
                },
                "summmonerId": {
                    "type": "string",
                    "uniqueItems": true
                },
                "accountId": {
                    "type": "string",
                    "uniqueItems": true
                }
            }
        },
        "Match": {
            "type": "object",
            "properties": {
                "puuid": {
                    "type": "string",
                    "uniqueItems": true
                },
                "championID": {
                    "type": "number"
                },
                "kills": {
                    "type": "number"
                },
                "deaths": {
                    "type": "number"
                },
                "KDA": {
                    "type": "number"
                },
                "time": {
                    "type": "number"
                }
            }
        },
        "Team": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "players": {
                    "items": {
                        "type": "string"
                    }
                }
            }
        }
    }
}